+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # :package: 0.1.0
 
+01. :gift: Added exception `dev.orne.test.rnd.GenerationException`
+01. :gift: Added exception `dev.orne.test.rnd.UnsupportedValueTypeException`
 01. :gift: Added interface `dev.orne.test.rnd.Generator`
 01. :gift: Added class `dev.orne.test.rnd.AbstractGenerator`
 01. :gift: Added interface `dev.orne.test.rnd.TypedGenerator`